Added YRTP in Windows, fixed compiling in VC++
git-svn-id: http://yate.null.ro/svn/yate/trunk@342 acf43c95-373e-0410-b603-e72c3f656dc1

+++ modules/ysipchan.cpp
@@ -814,7 +814,7 @@ SDPBody* YateSIPConnection::createSDP(const char* addr, const char* port, const 
     if (m_rtpSession)
 	++m_rtpVersion;
     else
-	m_rtpVersion = m_rtpSession = Time::now() / 10000000000ULL;
+	m_rtpVersion = m_rtpSession = (int)(Time::now() / (u_int32_t)10000000000);
     String owner;
     owner << "yate " << m_rtpSession << " " << m_rtpVersion << " IN IP4 " << addr;
     if (!port) {
